requirements.txt
# APOEMA - Core Dependencies
crewai>=0.1.0
crewai_files>=1.0.0

# LLM Provider
google-generativeai>=0.3.0

# Data Processing
pydantic>=2.0
python-dotenv>=1.0.0

# PDF Processing (Optional but recommended)
docling>=1.0.0
pypdfium2>=1.0.0

# Utilities
requests>=2.31.0
pyyaml>=6.0.0

# Development (Optional)
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
